+++ src/InformaticsGateway/Services/Http/SourceAeTitleController.cs
@@ -23,6 +23,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Monai.Deploy.InformaticsGateway.Api;
+using Monai.Deploy.InformaticsGateway.Common;
 using Monai.Deploy.InformaticsGateway.Configuration;
 using Monai.Deploy.InformaticsGateway.Logging;
 using Monai.Deploy.InformaticsGateway.Repositories;
@@ -91,6 +92,7 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
         [Consumes(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status409Conflict)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [Produces("application/json")]
         public async Task<ActionResult<string>> Create(SourceApplicationEntity item)
@@ -104,6 +106,10 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
                 await _repository.SaveChangesAsync().ConfigureAwait(false);
                 _logger.SourceApplicationEntityAdded(item.AeTitle, item.HostIp);
                 return CreatedAtAction(nameof(GetAeTitle), new { name = item.Name }, item);
+            }
+            catch (ObjectExistsException ex)
+            {
+                return Problem(title: "DICOM source already exists", statusCode: (int)System.Net.HttpStatusCode.Conflict, detail: ex.Message);
             }
             catch (ConfigurationException ex)
             {
@@ -148,11 +154,11 @@ namespace Monai.Deploy.InformaticsGateway.Services.Http
         {
             if (_repository.Any(p => p.Name.Equals(item.Name)))
             {
-                throw new ConfigurationException($"A DICOM source with the same name '{item.Name}' already exists.");
+                throw new ObjectExistsException($"A DICOM source with the same name '{item.Name}' already exists.");
             }
             if (_repository.Any(p => item.AeTitle.Equals(p.AeTitle) && item.HostIp.Equals(p.HostIp)))
             {
-                throw new ConfigurationException($"A DICOM source with the same AE Title '{item.AeTitle}' and host/IP address '{item.HostIp}' already exists.");
+                throw new ObjectExistsException($"A DICOM source with the same AE Title '{item.AeTitle}' and host/IP address '{item.HostIp}' already exists.");
             }
             if (!item.IsValid(out var validationErrors))
             {