+++ src/InformaticsGateway/Test/Services/Fhir/FhirJsonReaderTest.cs
@@ -16,11 +16,14 @@
 
 using System;
 using System.IO;
+using System.IO.Abstractions;
+using System.IO.Abstractions.TestingHelpers;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
 using Monai.Deploy.InformaticsGateway.Configuration;
 using Monai.Deploy.InformaticsGateway.Services.Fhir;
@@ -33,11 +36,17 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Fhir
     {
         private readonly InformaticsGatewayConfiguration _config;
         private readonly Mock<ILogger<FhirJsonReader>> _logger;
+        private readonly IOptions<InformaticsGatewayConfiguration> _options;
+        private readonly IFileSystem _fileSystem;
 
         public FhirJsonReaderTest()
         {
             _config = new InformaticsGatewayConfiguration();
             _logger = new Mock<ILogger<FhirJsonReader>>();
+            _options = Options.Create<InformaticsGatewayConfiguration>(new InformaticsGatewayConfiguration());
+            _fileSystem = new MockFileSystem();
+
+            _options.Value.Storage.TemporaryDataStorage = TemporaryDataStorageLocation.Memory;
         }
 
         [Fact]
@@ -46,7 +55,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Fhir
             var request = new Mock<HttpRequest>();
             var correlationId = Guid.NewGuid().ToString();
             var resourceType = "Patient";
-            var reader = new FhirJsonReader(_logger.Object);
+            var reader = new FhirJsonReader(_logger.Object, _options, _fileSystem);
 
             await Assert.ThrowsAsync<ArgumentNullException>(async () => await reader.GetContentAsync(null, null, null, null, CancellationToken.None));
             await Assert.ThrowsAsync<ArgumentNullException>(async () => await reader.GetContentAsync(request.Object, null, null, null, CancellationToken.None));
@@ -62,7 +71,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Fhir
             var correlationId = Guid.NewGuid().ToString();
             var resourceType = "Patient";
             var contentType = new MediaTypeHeaderValue(ContentTypes.ApplicationFhirJson);
-            var reader = new FhirJsonReader(_logger.Object);
+            var reader = new FhirJsonReader(_logger.Object, _options, _fileSystem);
 
             await Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
@@ -83,7 +92,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Fhir
             var correlationId = Guid.NewGuid().ToString();
             var resourceType = "Patient";
             var contentType = new MediaTypeHeaderValue(ContentTypes.ApplicationFhirJson);
-            var reader = new FhirJsonReader(_logger.Object);
+            var reader = new FhirJsonReader(_logger.Object, _options, _fileSystem);
 
             await Assert.ThrowsAnyAsync<JsonException>(async () =>
             {
@@ -106,7 +115,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Fhir
             var correlationId = Guid.NewGuid().ToString();
             var resourceType = "Patient";
             var contentType = new MediaTypeHeaderValue(ContentTypes.ApplicationFhirJson);
-            var reader = new FhirJsonReader(_logger.Object);
+            var reader = new FhirJsonReader(_logger.Object, _options, _fileSystem);
 
             var data = System.Text.Encoding.UTF8.GetBytes(xml);
             using var stream = new System.IO.MemoryStream();