+++ src/InformaticsGateway/Test/Services/Http/DestinationAeTitleControllerTest.cs
@@ -27,6 +27,7 @@ using Microsoft.Extensions.Logging;
 using Monai.Deploy.InformaticsGateway.Api;
 using Monai.Deploy.InformaticsGateway.Repositories;
 using Monai.Deploy.InformaticsGateway.Services.Http;
+using Monai.Deploy.InformaticsGateway.Services.Scu;
 using Moq;
 using xRetry;
 using Xunit;
@@ -38,11 +39,13 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Http
         private readonly DestinationAeTitleController _controller;
         private readonly Mock<ProblemDetailsFactory> _problemDetailsFactory;
         private readonly Mock<ILogger<DestinationAeTitleController>> _logger;
+        private readonly Mock<IScuQueue> _scuQueue;
         private readonly Mock<IInformaticsGatewayRepository<DestinationApplicationEntity>> _repository;
 
         public DestinationAeTitleControllerTest()
         {
             _logger = new Mock<ILogger<DestinationAeTitleController>>();
+            _scuQueue = new Mock<IScuQueue>();
 
             _problemDetailsFactory = new Mock<ProblemDetailsFactory>();
             _problemDetailsFactory.Setup(_ => _.CreateProblemDetails(
@@ -67,11 +70,14 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Http
 
             _repository = new Mock<IInformaticsGatewayRepository<DestinationApplicationEntity>>();
 
+            var controllerContext = new ControllerContext { HttpContext = new DefaultHttpContext() };
             _controller = new DestinationAeTitleController(
                  _logger.Object,
-                 _repository.Object)
+                 _repository.Object,
+                 _scuQueue.Object)
             {
-                ProblemDetailsFactory = _problemDetailsFactory.Object
+                ProblemDetailsFactory = _problemDetailsFactory.Object,
+                ControllerContext = controllerContext
             };
         }
 
@@ -172,6 +178,104 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Http
 
         #endregion GetAeTitle
 
+        #region C-Echo
+
+        [RetryFact(5, 250)]
+        public async Task GivenAnEmptyString_WhenCEchoIsCalled_Returns404()
+        {
+            var result = await _controller.CEcho(string.Empty);
+            var notFoundResult = result as NotFoundResult;
+            Assert.NotNull(notFoundResult);
+            Assert.Equal(StatusCodes.Status404NotFound, notFoundResult.StatusCode);
+        }
+
+        [RetryFact(5, 250)]
+        public async Task GivenADestinationName_WhenCEchoIsCalledAndEntityCannotBeFound_Returns404()
+        {
+            _repository.Setup(p => p.FindAsync(It.IsAny<string>())).ReturnsAsync(default(DestinationApplicationEntity));
+            var result = await _controller.CEcho("AET");
+            var notFoundResult = result as NotFoundResult;
+            Assert.NotNull(notFoundResult);
+            Assert.Equal(StatusCodes.Status404NotFound, notFoundResult.StatusCode);
+        }
+
+        [RetryFact(5, 250)]
+        public async Task GivenADestinationName_WhenCEchoIsCalledWithAnError_Returns502()
+        {
+            _repository.Setup(p => p.FindAsync(It.IsAny<string>()))
+                .ReturnsAsync(new DestinationApplicationEntity
+                {
+                    AeTitle = "AET",
+                    HostIp = "1.2.3.4",
+                    Port = 104,
+                    Name = "AET"
+                });
+            _scuQueue.Setup(p => p.Queue(It.IsAny<ScuWorkRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new ScuWorkResponse
+                {
+                    Status = ResponseStatus.Failure,
+                    Error = ResponseError.AssociationRejected,
+                    Message = "error"
+                });
+            var result = await _controller.CEcho("AET");
+            var objectResult = result as ObjectResult;
+            Assert.NotNull(objectResult);
+            var problemDetails = objectResult.Value as ProblemDetails;
+            Assert.NotNull(problemDetails);
+            Assert.Equal(StatusCodes.Status502BadGateway, problemDetails.Status);
+            Assert.Equal("C-ECHO Failure", problemDetails.Title);
+            Assert.Equal("error", problemDetails.Detail);
+        }
+
+        [RetryFact(5, 250)]
+        public async Task GivenADestinationName_WhenCEchoIsCalledWithUnhandledError_Returns500()
+        {
+            _repository.Setup(p => p.FindAsync(It.IsAny<string>()))
+                .ReturnsAsync(new DestinationApplicationEntity
+                {
+                    AeTitle = "AET",
+                    HostIp = "1.2.3.4",
+                    Port = 104,
+                    Name = "AET"
+                });
+            _scuQueue.Setup(p => p.Queue(It.IsAny<ScuWorkRequest>(), It.IsAny<CancellationToken>()))
+                .ThrowsAsync(new Exception("error"));
+            var result = await _controller.CEcho("AET");
+            var objectResult = result as ObjectResult;
+            Assert.NotNull(objectResult);
+            var problemDetails = objectResult.Value as ProblemDetails;
+            Assert.NotNull(problemDetails);
+            Assert.Equal(StatusCodes.Status500InternalServerError, problemDetails.Status);
+            Assert.Equal("Error performing C-ECHO", problemDetails.Title);
+            Assert.Equal("error", problemDetails.Detail);
+        }
+
+        [RetryFact(5, 250)]
+        public async Task GivenADestinationName_WhenCEchoIsCalledSuccessfully_Returns200()
+        {
+            _repository.Setup(p => p.FindAsync(It.IsAny<string>()))
+                .ReturnsAsync(new DestinationApplicationEntity
+                {
+                    AeTitle = "AET",
+                    HostIp = "1.2.3.4",
+                    Port = 104,
+                    Name = "AET"
+                });
+            _scuQueue.Setup(p => p.Queue(It.IsAny<ScuWorkRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(new ScuWorkResponse
+                {
+                    Status = ResponseStatus.Success,
+                    Error = ResponseError.None,
+                    Message = ""
+                });
+            var result = await _controller.CEcho("AET");
+            var okResult = result as OkResult;
+            Assert.NotNull(okResult);
+            Assert.Equal(StatusCodes.Status200OK, okResult.StatusCode);
+        }
+
+        #endregion C-Echo
+
         #region Create
 
         [RetryFact(5, 250, DisplayName = "GetAeTitle - Shall return problem on validation failure")]
@@ -192,9 +296,36 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Http
             Assert.NotNull(objectResult);
             var problem = objectResult.Value as ProblemDetails;
             Assert.NotNull(problem);
-            Assert.Equal("Validation error.", problem.Title);
+            Assert.Equal("Validation error", problem.Title);
             Assert.Equal($"'{aeTitle}' is not a valid AE Title (source: DestinationApplicationEntity).", problem.Detail);
             Assert.Equal((int)HttpStatusCode.BadRequest, problem.Status);
+        }
+
+        [RetryFact(5, 250, DisplayName = "Create - Shall return Conflict if entity already exists")]
+        public async Task Create_ShallReturnConflictIfEntityAlreadyExists()
+        {
+            var aeTitle = "AET";
+            var aeTitles = new DestinationApplicationEntity
+            {
+                Name = aeTitle,
+                AeTitle = aeTitle,
+                HostIp = "host",
+                Port = 1
+            };
+
+            _repository.Setup(p => p.Any(It.IsAny<Func<DestinationApplicationEntity, bool>>())).Returns(true);
+
+            var result = await _controller.Create(aeTitles);
+
+            var objectResult = result.Result as ObjectResult;
+            Assert.NotNull(objectResult);
+            var problem = objectResult.Value as ProblemDetails;
+            Assert.NotNull(problem);
+            Assert.Equal("DICOM destination already exists", problem.Title);
+            Assert.Equal("A DICOM destination with the same name 'AET' already exists.", problem.Detail);
+            Assert.Equal((int)HttpStatusCode.Conflict, problem.Status);
+
+            _repository.Verify(p => p.AddAsync(It.IsAny<DestinationApplicationEntity>(), It.IsAny<CancellationToken>()), Times.Never());
         }
 
         [RetryFact(5, 250, DisplayName = "Create - Shall return problem if failed to add")]
@@ -217,7 +348,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Http
             Assert.NotNull(objectResult);
             var problem = objectResult.Value as ProblemDetails;
             Assert.NotNull(problem);
-            Assert.Equal("Error adding new DICOM destination.", problem.Title);
+            Assert.Equal("Error adding new DICOM destination", problem.Title);
             Assert.Equal($"error", problem.Detail);
             Assert.Equal((int)HttpStatusCode.InternalServerError, problem.Status);