+++ src/InformaticsGateway/Test/Services/Storage/ObjectUploadServiceTest.cs
@@ -31,6 +31,7 @@ using Monai.Deploy.InformaticsGateway.Repositories;
 using Monai.Deploy.InformaticsGateway.Services.Storage;
 using Monai.Deploy.Storage.API;
 using Moq;
+using xRetry;
 using Xunit;
 
 namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
@@ -45,7 +46,6 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
         private readonly IObjectUploadQueue _uploadQueue;
         private readonly Mock<IStorageService> _storageService;
         private readonly Mock<IStorageMetadataWrapperRepository> _storageMetadataWrapperRepository;
-
         private readonly CancellationTokenSource _cancellationTokenSource;
         private readonly ServiceProvider _serviceProvider;
         private readonly Mock<IServiceScope> _serviceScope;
@@ -123,8 +123,8 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
             _storageService.Verify(p => p.PutObjectAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Stream>(), It.IsAny<long>(), It.IsAny<string>(), It.IsAny<Dictionary<string, string>>(), It.IsAny<CancellationToken>()), Times.Exactly(2));
         }
 
-        [Fact]
-        public void GivenAFhirFileStorageMetadata_WhenQueuedForUpload_ExpectSingleFileToBeUploaded()
+        [RetryFact]
+        public async Task GivenAFhirFileStorageMetadata_WhenQueuedForUpload_ExpectSingleFileToBeUploaded()
         {
             var countdownEvent = new CountdownEvent(1);
             _storageService.Setup(p => p.PutObjectAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Stream>(), It.IsAny<long>(), It.IsAny<string>(), It.IsAny<Dictionary<string, string>>(), It.IsAny<CancellationToken>()))
@@ -137,7 +137,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
 
             Assert.Equal(ServiceStatus.Running, svc.Status);
 
-            var file = GenerateFhirFileStorageMetadata();
+            var file = await GenerateFhirFileStorageMetadata();
             _uploadQueue.Queue(file);
 
             Assert.True(countdownEvent.Wait(TimeSpan.FromSeconds(3)));
@@ -145,11 +145,11 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
             _storageService.Verify(p => p.PutObjectAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Stream>(), It.IsAny<long>(), It.IsAny<string>(), It.IsAny<Dictionary<string, string>>(), It.IsAny<CancellationToken>()), Times.Once());
         }
 
-        private FhirFileStorageMetadata GenerateFhirFileStorageMetadata()
+        private async Task<FhirFileStorageMetadata> GenerateFhirFileStorageMetadata()
         {
             var file = new FhirFileStorageMetadata(Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), FhirStorageFormat.Json);
 
-            file.SetDataStream("[]");
+            await file.SetDataStream("[]", TemporaryDataStorageLocation.Memory);
             return file;
         }
 
@@ -169,7 +169,7 @@ namespace Monai.Deploy.InformaticsGateway.Test.Services.Storage
                 { DicomTag.SOPClassUID, DicomUID.SecondaryCaptureImageStorage.UID }
             };
             var dicomFile = new DicomFile(dataset);
-            await file.SetDataStreams(dicomFile, "[]");
+            await file.SetDataStreams(dicomFile, "[]", TemporaryDataStorageLocation.Memory);
             return file;
         }
     }