+++ docs/changelog.md
@@ -23,6 +23,27 @@
 - Adds a basic [FHIR service](api/rest/config.md) to accept any versions of FHIR.
 - Updates [Health Check API](api/rest/health.md) to replace `/health/live` and `/health/ready` APIs with `/health` API.
 
+## 0.3.1
+
+[GitHub Milestone 0.3.2](https://github.com/Project-MONAI/monai-deploy-informatics-gateway/milestone/8)
+
+- New [C-ECHO API](api/rest/config.md) to perform a DICOM C-ECHO to a configured DICOM destination.
+
+## 0.3.1
+
+[GitHub Milestone 0.3.1](https://github.com/Project-MONAI/monai-deploy-informatics-gateway/milestone/7)
+
+- The SCU AE Title is now uppercase MONAISCU.
+- Update fo-dicom to 5.0.3
+- Defaults temporary storage to use disk with ability to switch to memory.
+
+## 0.3.0
+
+[GitHub Milestone 0.3.0](https://github.com/Project-MONAI/monai-deploy-informatics-gateway/milestone/3)
+
+- Adds a basic [FHIR service](api/rest/fhir.md) to accept any versions of FHIR.
+- Updates [Health Check API](api/rest/health.md) to replace `/health/live` and `/health/ready` APIs with `/health` API.
+
 ## 0.2.0
 
 [GitHub Milestone 0.2.0](https://github.com/Project-MONAI/monai-deploy-informatics-gateway/milestone/6)
@@ -36,7 +57,7 @@
   any pending files. Files that were successfully uploaded to the temporary location (`storage>temporary`) in the 
   bucket (`storage>temporaryBucketName`) are then moved to the payload bucket (`storage>bucketName`) before sending a workflow request.  
 - Breaking changes in the storage configuration to allow dynamic key-value pairs.
-- Breaking changes to enable dynamic loadig of the storage & the messaging libraries.
+- Breaking changes to enable dynamic loading of the storage & the messaging libraries.
 
 ## 0.1.1